% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denguedat.R
\docType{data}
\name{denguedat}
\alias{denguedat}
\title{denguedat: Dengue fever reporting data from Puerto Rico}
\format{A data frame.}
\usage{
data(denguedat)
}
\description{
Surveillance data from CDC Division of Vector-Borne Diseases.
1990-2010 case reporting data included.
The first column, \code{onset_week}, indicates the week of symptom onset (case occurrence).
The second column, \code{report_week}, indicates the week of case report.
}
\examples{
data(denguedat)
NobBS(denguedat, as.Date("1990-10-01"),units="1 week",onset_date="onset_week",report_date="report_week")
}
\keyword{datasets,}
\keyword{dengue}
\keyword{fever}
